backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
Authlib==1.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==38.0.1
dj-database-url==0.5.0
Django==3.2
django-autoslug==1.9.8
django-bootstrap==0.2.4
django-ckeditor==6.1.0
django-cors-headers==3.13.0
django-crispy-forms==1.12.0
django-dotenv==1.4.2
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.13.1
dnspython==2.2.1
ecdsa==0.18.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.31
pyasn1==0.4.8
pycparser==2.21
pystache==0.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
requests==2.26.0
rsa==4.9
six==1.16.0
sqlparse==0.4.2
static==1.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
